+++ iOS/PrivateFrameworks/SpotlightServices.framework/PRSRankingItemRelativeFeatureContext.h
@@ -10,13 +10,13 @@
 
 @interface PRSRankingItemRelativeFeatureContext : NSObject
 {
-    unsigned long long _relRankFeature;
-    unsigned long long _absRankFeature;
+    unsigned short _relRankFeature;
+    unsigned short _absRankFeature;
     NSPredicate *_predicate;
     CDUnknownBlockType _comparator;
 }
 
-+ (void)_associatedPRSL2FeatureFromRelativeFeature:(unsigned long long)arg1 absRankFeatureOut:(unsigned long long *)arg2 relRankFeatureOut:(unsigned long long *)arg3;
++ (void)_associatedPRSL2FeatureFromRelativeFeature:(unsigned long long)arg1 absRankFeatureOut:(unsigned short *)arg2 relRankFeatureOut:(unsigned short *)arg3;
 + (id)_predicateForRelativeFeature:(unsigned long long)arg1;
 + (CDUnknownBlockType)_resultComparatorForRelativeFeature:(unsigned long long)arg1;
 + (id)_relativeContextsForFeatures:(id)arg1;
@@ -24,8 +24,8 @@
 + (id)relativeContextsForBundle:(id)arg1;
 @property(copy, nonatomic) CDUnknownBlockType comparator; // @synthesize comparator=_comparator;
 @property(retain, nonatomic) NSPredicate *predicate; // @synthesize predicate=_predicate;
-@property(nonatomic) unsigned long long absRankFeature; // @synthesize absRankFeature=_absRankFeature;
-@property(nonatomic) unsigned long long relRankFeature; // @synthesize relRankFeature=_relRankFeature;
+@property(nonatomic) unsigned short absRankFeature; // @synthesize absRankFeature=_absRankFeature;
+@property(nonatomic) unsigned short relRankFeature; // @synthesize relRankFeature=_relRankFeature;
 - (void).cxx_destruct;
 - (id)initWithFeature:(unsigned long long)arg1;
