+++ macOS/PrivateFrameworks/AVConference.framework/VCRateControlOWRDEstimator.h
@@ -27,7 +27,7 @@ __attribute__((visibility("hidden")))
 @property(readonly, nonatomic) double owrd; // @synthesize owrd=_owrd;
 - (void)calculateOWRDWithSendTime:(double)arg1 receiveTime:(double)arg2;
 - (double)relativeReceiveTimeWithTimestamp:(unsigned short)arg1 timestampRate:(unsigned short)arg2;
-- (double)relativeSendTimeWithTimestamp:(unsigned int)arg1 timestampRate:(unsigned int)arg2;
+- (double)relativeSendTimeWithTimestamp:(unsigned short)arg1 timestampRate:(unsigned short)arg2;
 
 @end
