+++ macOS/PrivateFrameworks/TimeSync.framework/TSgPTPMachPort.h
@@ -13,7 +13,7 @@
 + (id)diagnosticDescriptionForService:(unsigned int)arg1 withIndent:(id)arg2;
 + (id)iokitMatchingDictionaryForClockIdentifier:(unsigned long long)arg1;
 - (void)dealloc;
-- (long long)portType;
+- (int)portType;
 - (id)initWithMatchingDictionary:(id)arg1;
 - (id)initWithService:(unsigned int)arg1;
